fix:tests

+++ src/math/fxpt.h
@@ -7,9 +7,9 @@
 
 namespace math {
 	class fxpt {
-		std::uint32_t raw;
+		std::int32_t raw;
 	public:
-		static constexpr std::uint32_t raw_bias = 65536;
+		static constexpr std::int32_t raw_bias = 65536;
 		static constexpr float         bias     = static_cast<float>(raw_bias);
 
 		__attribute__((always_inline)) constexpr inline fxpt(float const f = 0);
@@ -50,12 +50,12 @@ namespace math {
 
 __attribute__((always_inline)) constexpr inline
 math::fxpt::fxpt(float const f) :
-	raw(static_cast<std::uint32_t>(static_cast<std::int32_t>(std::round(f * math::fxpt::bias))))
+	raw(static_cast<std::int32_t>(std::round(f * math::fxpt::bias)))
 { }
 
 __attribute__((always_inline)) constexpr inline float
 math::fxpt::to_float() const {
-	return static_cast<float>(static_cast<std::int32_t>(raw)) / math::fxpt::bias;
+	return static_cast<float>(raw) / math::fxpt::bias;
 }
 
 __attribute__((always_inline)) constexpr inline
@@ -70,12 +70,12 @@ math::fxpt::operator float() const {
 
 __attribute__((always_inline)) constexpr inline
 math::fxpt::operator double() const {
-	return static_cast<double>(static_cast<std::int32_t>(raw)) / math::fxpt::bias;
+	return static_cast<double>(raw) / math::fxpt::bias;
 }
 
 __attribute__((always_inline)) constexpr inline
 math::fxpt::operator int() const {
-	return static_cast<int>(static_cast<std::int32_t>(raw) / static_cast<std::int32_t>(math::fxpt::raw_bias));
+	return static_cast<int>(raw / math::fxpt::raw_bias);
 }
 
 __attribute__((always_inline)) constexpr inline math::fxpt
@@ -95,17 +95,17 @@ math::fxpt::operator-(math::fxpt const other) const {
 __attribute__((always_inline)) constexpr inline math::fxpt
 math::fxpt::operator-() const {
 	math::fxpt result{};
-	result.raw = static_cast<std::uint32_t>(-static_cast<std::int32_t>(raw));
+	result.raw = -raw;
 	return result;
 }
 
 __attribute__((always_inline)) constexpr inline math::fxpt
 math::fxpt::operator*(math::fxpt const other) const {
-	std::int64_t a = static_cast<std::int64_t>(static_cast<std::int32_t>(raw));
-	std::int64_t b = static_cast<std::int64_t>(static_cast<std::int32_t>(other.raw));
+	std::int64_t a = static_cast<std::int64_t>(raw);
+	std::int64_t b = static_cast<std::int64_t>(other.raw);
 	std::int64_t product = a * b;
 	math::fxpt result{};
-	result.raw = static_cast<std::uint32_t>((product + (math::fxpt::raw_bias/2)) / math::fxpt::raw_bias);
+	result.raw = (product + (math::fxpt::raw_bias/2)) / math::fxpt::raw_bias;
 	return result;
 }
 
@@ -114,12 +114,12 @@ math::fxpt::operator/(math::fxpt const other) const {
 	if(other.raw == 0) {
 		return math::fxpt(0.0f);
 	}
-	std::int64_t a = static_cast<std::int64_t>(static_cast<std::int32_t>(raw));
-	std::int64_t b = static_cast<std::int64_t>(static_cast<std::int32_t>(other.raw));
+	std::int64_t a = static_cast<std::int64_t>(raw);
+	std::int64_t b = static_cast<std::int64_t>(other.raw);
 	std::int64_t divident = a * math::fxpt::raw_bias;
 	std::int64_t quotient = divident / b;
 	math::fxpt result{};
-	result.raw = static_cast<std::uint32_t>(quotient);
+	result.raw = quotient;
 	return result;
 }
 
@@ -159,22 +159,22 @@ math::fxpt::operator!=(math::fxpt const other) const {
 
 __attribute__((always_inline)) constexpr inline bool
 math::fxpt::operator<(math::fxpt const other) const {
-	return static_cast<std::int32_t>(raw) < static_cast<std::int32_t>(other.raw);
+	return raw < other.raw;
 }
 
 __attribute__((always_inline)) constexpr inline bool
 math::fxpt::operator>(math::fxpt const other) const {
-	return static_cast<std::int32_t>(raw) > static_cast<std::int32_t>(other.raw);
+	return raw > other.raw;
 }
 
 __attribute__((always_inline)) constexpr inline bool
 math::fxpt::operator<=(math::fxpt const other) const {
-	return static_cast<std::int32_t>(raw) <= static_cast<std::int32_t>(other.raw);
+	return raw <= other.raw;
 }
 
 __attribute__((always_inline)) constexpr inline bool
 math::fxpt::operator>=(math::fxpt const other) const {
-	return static_cast<std::int32_t>(raw) >= static_cast<std::int32_t>(other.raw);
+	return raw >= other.raw;
 }
 
 constexpr math::fxpt
